+++ src/app/screens/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "@fontsource/roboto";
 
-import Router from "@app/Router";
+import Router from "@components/Router";
 
 const App = () => <Router />;
