+++ src/server/app.ts
@@ -2,9 +2,8 @@ import express from "express";
 import session from "express-session";
 import i18next from "i18next-http-middleware";
 import methodOverride from "method-override";
-import passport from "passport";
 
-import "@server/auth";
+import passport from "@server/auth";
 import i18n from "i18n";
 import routes from "routes";
